+++ src/main/java/racingcar/Application.java
@@ -1,7 +1,11 @@
 package racingcar;
 
+import java.util.Arrays;
+
+import racingcar.controller.GameController;
+
 public class Application {
 	public static void main(String[] args) {
-		GameHandler.startRace();
+		GameController.startRace();
 	}
 }
