+++ src/main/java/racingcar/controller/GameController.java
@@ -1,12 +1,15 @@
-package racingcar;
+package racingcar.controller;
 
-public class GameHandler {
+import racingcar.domain.CarList;
+import racingcar.view.Input;
+import racingcar.view.Output;
+
+public class GameController {
 	public static void startRace() {
 		String[] carNames = Input.carName();
 		int tryNumber = Input.tryNumber();
 
 		CarList cars = new CarList(carNames);
-
 		for (int i = 0; i < tryNumber; i++) {
 			cars.moveAll();
 			Output.movementState(cars);
